minor: reorg preprocessing debug logging and refactor raw directory handling
